refactor(network): make continuous network discovery into a single module
