+++ ant-networking/src/lib.rs
@@ -9,7 +9,6 @@
 #[macro_use]
 extern crate tracing;
 
-mod bootstrap;
 mod circular_vec;
 mod cmd;
 mod config;
